More work

+++ opt/src/commonMain/kotlin/ir/platform/x64/codegen/impl/MemcpyCodegen.kt
@@ -17,15 +17,15 @@ internal class MemcpyCodegen(val length: UnsignedIntegerConstant, val asm: Assem
 
     override fun rr(dst: GPRegister, src: GPRegister) {
         // Not so efficient, but simple
-        val iterations8 = length.value() / POINTER_SIZE.toUInt()
-        val remains8 = length.value() % POINTER_SIZE.toUInt()
+        val iterations8 = length.value() / POINTER_SIZE
+        val remains8 = length.value() % POINTER_SIZE
         for (i in 0 until iterations8.toInt()) {
             asm.mov(POINTER_SIZE, Address.from(src, i * POINTER_SIZE), temp1)
             asm.mov(POINTER_SIZE, temp1, Address.from(dst, i * POINTER_SIZE))
         }
         val base = iterations8.toInt() * POINTER_SIZE
-        val iterations4 = remains8 / WORD_SIZE.toUInt()
-        val remains4 = remains8 % WORD_SIZE.toUInt()
+        val iterations4 = remains8 / WORD_SIZE
+        val remains4 = remains8 % WORD_SIZE
         for (i in 0 until iterations4.toInt()) {
             asm.mov(WORD_SIZE, Address.from(src, base + i * WORD_SIZE), temp1)
             asm.mov(WORD_SIZE, temp1, Address.from(dst, base + i * WORD_SIZE))
